refactor(web): chat の hook/context をフォルダ規約に沿って整理
dashboard の慣習（contexts/ に context+Provider+フック、hooks/ に各 hook）に揃える。
- ChatProvider.tsx + useChatContext.ts(→ChatContext.ts) を contexts/ へ
- useThreadManager.ts を hooks/ へ
- feedback-helpers.ts は feature util として直下に残す
- vitest.config.ts の coverage exclude を移動後パスに更新し、削除済み
  FeedbackContext / FeedbackModalWrapper の stale エントリを除去
